Extract shared JWT token decoding to oauth2_common
- Move decode_id_token logic to oauth2_common as generic function
- Move base64_decode helper to oauth2_common
- Update Apple and Google modules to use shared implementation
- Eliminate 28 lines of duplicated JWT decoding code
