+++ tee-worker/omni-executor/heima/identity-verification/src/web2/oauth2_common.rs
@@ -1,4 +1,6 @@
 use crate::helpers;
+use base64::prelude::{Engine, BASE64_URL_SAFE_NO_PAD};
+use serde::de::DeserializeOwned;
 use url::Url;
 
 pub struct AuthorizeData {
@@ -31,4 +33,19 @@ pub fn get_authorize_data(
 	}
 
 	AuthorizeData { authorize_url: authorize_url.into(), state }
+}
+
+pub fn decode_id_token<T: DeserializeOwned>(token: &str) -> Result<T, &'static str> {
+	let parts: Vec<&str> = token.split('.').collect();
+	if parts.len() != 3 {
+		return Err("Invalid token format");
+	}
+	let payload = base64_decode(parts[1])?;
+	serde_json::from_str(&payload).map_err(|_| "Failed to parse claims")
+}
+
+fn base64_decode(input: &str) -> Result<String, &'static str> {
+	let decoded = &BASE64_URL_SAFE_NO_PAD.decode(input).map_err(|_| "Failed to decode base64")?;
+
+	Ok(String::from_utf8_lossy(decoded).to_string())
 }
